Create code
Module -11 Debugging and defensive programming in R

In order to complete the assignment, I must run the provided code, determine whether it succeeds or fails, and then troubleshoot any issues.

Given code:

tukey_multiple <- function(x) {
   outliers <- array(TRUE,dim=dim(x))
   for (j in 1:ncol(x))
    {
    outliers[,j] <- outliers[,j] && tukey.outlier(x[,j])
    }
outlier.vec <- vector(length=nrow(x))
    for (i in 1:nrow(x))
    { outlier.vec[i] <- all(outliers[i,]) } return(outlier.vec) }

Outcomes I received:



Error: unexpected symbol in:
"    for (i in 1:nrow(x))
    { outlier.vec[i] <- all(outliers[i,]) } return"


The error can be seen as a caret under the letter r in the word "return."

You must return the value or a string in a distinct line since the return() function cannot be written in the same line as the curly brackets. It should seem as follows:

tukey_multiple <- function(x) {
  outliers <- array(TRUE, dim = dim(x))
  for (j in 1:ncol(x)) 
    {
    outliers[, j] <- outliers[, j] && tukey.outlier(x[, j])
    }
  outlier.vec <- vector(length = nrow(x))
  for (i in 1:nrow(x))
    { outlier.vec[i] <- all(outliers[i,]) }
  return(outlier.vec) }
and here with this It get successful because I ran this revised code and 
had no issues.